requirements.txt
requests>=2.28.0
openai>=0.27.0
python-dateutil>=2.8.2
aiohttp>=3.8.0
pyyaml>=6.0.0
markdown>=3.4.0 